Add new folder structure in ProjectFolder
ProjectFolder is now default project folder. Super Folder is used for foreign program code and report and literature.

+++ Helpers/IO.py
@@ -1,8 +1,7 @@
+import os
 from mido import MidiFile
 import numpy as np
-
-import wave	# wave module in python
-
+import wave  # wave module in python
 from scipy.io import wavfile
 
 """
@@ -11,14 +10,21 @@ output : numpy array of type string, containing all MIDI data
 
 needs : mido package  : pip install mido 
  
-Todo : check for valid input file
-       partial read 
+Todo : partial read 
 """
-def readMIDI (f):	
+
+
+def readMIDI(f):
+    if not ('.mid' in f):
+        raise ValueError(f + ' is not a valid file path of type .mid')
+
+    if not (os.path.isfile(f)):
+        raise FileNotFoundError
+
     mid = MidiFile(f)
     ostr = []
 
-    for i, track in enumerate (mid.tracks):  
+    for i, track in enumerate(mid.tracks):
         ostr.append('Track {}: {}\n'.format(i, track.name))
         for msg in track:
             ostr.append(msg)
@@ -33,17 +39,24 @@ output : text file containing all MIDI data
 
 needs : mido package  : pip install mido 
  
-Todo : check for valid input file
-       partial read 
+Todo : partial read 
 """
-def MIDI2file (ifi, ofo=''):	
+
+
+def MIDI2file(ifi, ofo=''):
+    if not ('.mid' in ifi):
+        raise ValueError(ifi + ' is not a valid file path of type .mid')
+
+    if not (os.path.isfile(ifi)):
+        raise FileNotFoundError
+
     mid = MidiFile(ifi)
     ostr = []
 
     if (ofo == ''):
         ofo = ifi.replace(".mid", "_MIDI.txt")
     sw = open(ofo, "w")
-    for i, track in enumerate (mid.tracks):  
+    for i, track in enumerate(mid.tracks):
         sw.write('Track {}: {}\n'.format(i, track.name))
         for msg in track:
             sw.write(str(msg))
@@ -58,14 +71,16 @@ needs : wave module in python
 Todo : check for valid input file
        read n frames (partial read)
 """
+
+
 def readWAV(f):
     sr = wave.openfp(f, "r")
     nchannels, sampwidth, framerate, nframes, comptype, compname = sr.getparams()  # for future use, seems not available in scipy.io.wavfile
 
-    fs, data = wavfile.read(f) # could also use sr.readframes(n) :  sr.readframes doesn't seem to deliver left / right channels 
+    fs, data = wavfile.read(
+        f)  # could also use sr.readframes(n) :  sr.readframes doesn't seem to deliver left / right channels
 
     return data
-
 
 
 """
@@ -78,6 +93,8 @@ Todo : check for valid input file
 not very useful because output file size might be huge
 better output into an npy, a pcl (?) or some compressed format
 """
+
+
 def WAV2file(ifi, ofo=''):
     sr = wave.openfp(ifi, "r")
     nchannels, sampwidth, framerate, nframes, comptype, compname = sr.getparams()  # for future use, seems not available in scipy.io.wavfile
@@ -85,9 +102,10 @@ def WAV2file(ifi, ofo=''):
     if (ofo == ''):
         ofo = ifi.replace(".wav", "_np.txt")
 
-    fs, data = wavfile.read(ifi) # could also use sr.readframes(n) :  sr.readframes doesn't seem to deliver left / right channels 
+    fs, data = wavfile.read(
+        ifi)  # could also use sr.readframes(n) :  sr.readframes doesn't seem to deliver left / right channels
 
-    wf = open(ofo, "w")    # open a text file, python wave module 
+    wf = open(ofo, "w")  # open a text file, python wave module
 
     for i in range(len(data)):
         print(data, file=wf)
